311 Added support for maintenance alerts
- Created api to support maintenance alerts

Created struct to handle the notifications and created seperate
struct to handle the update process of the notification. Since we do not
need a lot of the struct values, we just created a new struct in leu of
reusing the older struct.

Created apilayer code to support displaying a list of maintenance alerts
that are not read for the user. Added sql code to retrieve the alerts that
are made by the collaborators of the alerts table. This collaborators are
populated from the maintenance plan table. So in theory, if the maintenance
plan has more collaborators the alert is shared across all of them.

Any one can mark the alert as complete. The userID of the person who marked
the alert complete is displayed in the updated by section of the table and
is inclusive of the timestamp of the performed action. If a new alert is created
by the system then the table shall read the label - system.

- Created sql functions to support maintenance alerts table

Added sql table to house the alerts for each maintenance plan. There should be
only one alert per maintenance plan. If the user changes the maintenance plan,
then the alert is re-populated if necessary ( if the conditions for the alert )
match.

Added pg cron extension and updated the docker container to house the extension.
By doing this, we are able to use the pg_cron extension and fire up the cron
scheduler to update the table at a selected time every day to build new alerts.

Added trigger function to update the alerts table after change is occured
on the maintenance plan table. This is required because if a user changes
a maintenance plan to be within the 7 days period of due date, then the
trigger function will update the alert table.

- Updated Redux and Saga to handle the new alerts api

Added fetch functions to retrieve the notifications and update functions
to update the notifications alerts from the redux action. The redux instead
does a opmistic update to the code, since the db only returns 200 ok if the
put is successful.

- Created new UI components to support the notifications popover content

Added new toolbar icons and popover content to display the notifications.
Clicking on any of the selected toolbar notification will attempt to mark the
notification as is_read or complete. This in turn will remove the notification
next time the page is viewed.

Closes #311

support ci changes for pg_cron

support ci changes for pg_cron

support ci changes for pg_cron

support ci changes for pg_cron

attempt to fix github action for custom image

Add logs to docker pipeline

Add logs to docker pipeline

add password

add password