313 - Update bug squash
Uplifted the snackbar to the main component of the app.
This allows us to bring snackbar messages when the Sign up component
works / does not work as expected.

Update height of the map in selected category / maintenance plan
Added map height to match the height of the card component. In the
small view mode the image is removed since the height is not present.

Remove disabled function

At first, it made sense to add the isDisabled() fn but later on realized
that there is very little we would be able to test from the ui. Since the
file type for the csv is long and complecated, we didn't want to introduce
this aspect in the code. If the user adds some other file and does not see
data then, that should most likely be a user error since the teimplate is
already built out. Plus we also wanted to check for documents from 2003
till 2007 and documents from 2007 and out.

Seperate storage locations to seed

Addressed fix for reuploading same items

Fixed the bug where adding the same item would not crash the page anymore.
Previously, it used to break the system, while re-adding the same item
but that was because of a stale state of selectedIDList. When the audience
member removes the item from the category, we reset the selectedIDList state.
This gives us a clean state to pick off from.

Updated category saga and redux actions to use selectedIDList instead of
using rowSelected state.

Closes #313 #278