WIP - Attempted to work with react tour
In our ongoing efforts to integrate the application with react tour,
we are trying to bridge various huddles that need to be ironed out. Couple
of issues that we are getting stuck at is that the tour needs to be
interactive but unfortunately, it is not interactive. Eg, the tour must
teach the user to add an asset, but state management is not working as
expected with react tour. we have thought of trying to diversify this
strategy.