313 Update minor bug fixes
Fixed an issue with selected asset with image.

Uploading an image in the selected asset page would not display the
image on the fly instead we would have to refresh the page. Fixed the issue
by reusing the redux fn uploadAndRefreshDataSuccess so that we can update
the selected image from the system on the fly.

Fixed an issue with collaborators not updating for categories and plans.

Collaborators for categories and / or maintenance plans were not being updated
since they were not attached properly to the redux action of category item after
the split between categories and category item. Hindsight 20/20 maybe we should have
just left those two alone since they were related but nevertheless the issue is resolved.
Simple fix was to just add the proper dispatch actions and on the slice, we just reset
the value of the image if new image was uploaded so that would force the dispatch action
to call the new image.

We found some discripancies with the way categories and maintenance plans are updated, since
the latter one uses image in its payload. we probably need to investigate that just to make
sure everything is correctly functioning and working as intended.

Closes #313