Strengthen validation logic to sign up
The validation logic for email field during signup was not working as
expected and it would just immediately validate a field even though the field did not exist.

Applied bandaid to that scenario by just ensuring that the dispatch action would be called only if there was a valid
email address or email address of at the very least 3 characters.