315 Add ability to send email notifications
Created new api endpoints

- Created api endpoints in api.go to reset email verification token if the token
does not get sent to the user. Also created new endpoint to handle verification
service of the selected token to ensure that it is a valid jwt.

- Added new handler service for the reset email token handler and created placeholder for the
verify email address handler functions. Updated swagger and wrote new definitions for the
endpoint as well.

Created UserService.go file to handle service related changes for Userdb handler.

- Created new file to properly handle the creation of user data and email notification service.
Currently, if the notification service fails, we silently fail and this is done so that
the application is not closely tied with email notification service since the service might not
be up at all times. To support both use cases, error with the email_notification_service is
logged but does not fail the user in the login process.

- Added utils package for common utility stuffs

- Created function to validate Jwt token if found.

Updated dev.env file

- Updated dev.env file to support SendGrid email notification service and added a feature flag that
can be used to toggle on / off the feature request of email notification serivce. If the feature flag
is turned off, then the api call is not made, and hence the email notification service will not work.

- Added the feature flags and other env variable changes to prod.env as well.

Added tech debt changes

- Updated main.go file to be renamed as api.go for user readability. Also updated
api.go file to better format the validation logic. renamed functions to provide
more meaningful output. Removed unused functions or functions that do not need to
be seperated out on its own.

- Updated packages and added ciri-stormrider pacakge to handle JWT authentication.
Instead of trying to re-write the logic ourselves, decided to create a package that
could be used in all earmuffjam apps which is a big bonus to us as well for re-usability
and scalability. Currently pacakge is under v.1.0.3 and can be updated at a later date
if newer / better features are requested / required.

- Performed cleanup on the Userdb file since a lot of work for the jwt can be re-purposed and
can be used from the package itself.

- Removed unused structs and functions from the User.go model layer. Removed WebsocketHandler code
as well since it is not being used.

Closes #315