Added bugfix for invalidation of cookie
Added bugfix where the user credentials are not properly discarded
if the token cannot be persisted in the db. Previously, the token
would be sent out instead of the masked ID, and hence the token would
be exposed.

Fixed this issue by ensuring the ID being added to the license key was
a valid and masked ID.