Add unit test and revamp handlers
Added unit test for service layer code

- We wish to reach a greater threshold of unit tests, but there are some
limitations that we cannot pass at this time with the limited amount of
time in the ticket. Since, we do not have a way to mock and spy on existing
code, we do not have a way to completely test all aspects of code. This
really shows when we are trying to test functions within a function. However,
we do have a decent amount of code coverage and will continue to work on
getting other parts of the code also covered thoroughly.

Updated main.sh file to support new test location

- Updated the main.sh file so that we can test all backend code instead of
just handler code.