Update backend support for email validation
Updated jwt package into new version 1.0.5

- This version supports parsing of the jwt token easily, hence we
moved to the updated version for the application.

Added Db support for verifying user

- Added code to support verification of the user in the db. However,
there is more work that is remaining on this front. We also need to
update the db layer for new field that can be used to verify the user.

- Added handler code to support user verification and also updated
the service layer for user validation and validation.

Tech debt changes

- Fixed unit test that was erorring out after moving to new version
for the jwt pacakge ciri-stormrider