Update DB and Web Layer for Email Service
Created new sql file to handle verified user

- Added new sql file to support the verification of a selected user.
Also updated test to support verified users. Currently, both test users
are labelled as verified, but when the MVP is rolled out, and new changes
are sought out for the implications of verified user, we can either
create a new unverified user or label one of the other users as an unverified
user.

- Updated api layer code to support revalidation of email for each user

The verified user route is a little complex since there is multiple avenues
of approach for this situation. Currently, users are able revalidate once
if the email didn't reach them but we want to ensure that the user has the
correct token and authorization so that they can perform the email notification
service again.

Updated model struct and test cases around user service.

- Updated UI to support email re-validation

The ui part for this request is a bit tricky. To prevent the users from
pressing the button again and again, we added a timeout for 60 seconds before
the user can send the email again. However, if the user refreshes the page, the
timer is invalidated. Hence, we needed a better solution. For now, we
simply store the key in the local storage and the next time the user logs
in they are able to request a re-validation again. Currently, there is
no visible / tangible issue with user invalidation. So for MVP we are going
to keep it at this.