327 feature/email service (#330)
* 315 Add ability to send email notifications

Created new api endpoints

- Created api endpoints in api.go to reset email verification token if the token
does not get sent to the user. Also created new endpoint to handle verification
service of the selected token to ensure that it is a valid jwt.

- Added new handler service for the reset email token handler and created placeholder for the
verify email address handler functions. Updated swagger and wrote new definitions for the
endpoint as well.

Created UserService.go file to handle service related changes for Userdb handler.

- Created new file to properly handle the creation of user data and email notification service.
Currently, if the notification service fails, we silently fail and this is done so that
the application is not closely tied with email notification service since the service might not
be up at all times. To support both use cases, error with the email_notification_service is
logged but does not fail the user in the login process.

- Added utils package for common utility stuffs

- Created function to validate Jwt token if found.

Updated dev.env file

- Updated dev.env file to support SendGrid email notification service and added a feature flag that
can be used to toggle on / off the feature request of email notification serivce. If the feature flag
is turned off, then the api call is not made, and hence the email notification service will not work.

- Added the feature flags and other env variable changes to prod.env as well.

Added tech debt changes

- Updated main.go file to be renamed as api.go for user readability. Also updated
api.go file to better format the validation logic. renamed functions to provide
more meaningful output. Removed unused functions or functions that do not need to
be seperated out on its own.

- Updated packages and added ciri-stormrider pacakge to handle JWT authentication.
Instead of trying to re-write the logic ourselves, decided to create a package that
could be used in all earmuffjam apps which is a big bonus to us as well for re-usability
and scalability. Currently pacakge is under v.1.0.3 and can be updated at a later date
if newer / better features are requested / required.

- Performed cleanup on the Userdb file since a lot of work for the jwt can be re-purposed and
can be used from the package itself.

- Removed unused structs and functions from the User.go model layer. Removed WebsocketHandler code
as well since it is not being used.

Closes #315

* Added bugfix for invalidation of cookie

Added bugfix where the user credentials are not properly discarded
if the token cannot be persisted in the db. Previously, the token
would be sent out instead of the masked ID, and hence the token would
be exposed.

Fixed this issue by ensuring the ID being added to the license key was
a valid and masked ID.

* Added comments and split features out

Added comments to required functions, updated log statements and
created service folder locations so that db does not get too cluttered
with the backend features.

* Add unit test and revamp handlers

Added unit test for service layer code

- We wish to reach a greater threshold of unit tests, but there are some
limitations that we cannot pass at this time with the limited amount of
time in the ticket. Since, we do not have a way to mock and spy on existing
code, we do not have a way to completely test all aspects of code. This
really shows when we are trying to test functions within a function. However,
we do have a decent amount of code coverage and will continue to work on
getting other parts of the code also covered thoroughly.

Updated main.sh file to support new test location

- Updated the main.sh file so that we can test all backend code instead of
just handler code.

* Update backend support for email validation

Updated jwt package into new version 1.0.5

- This version supports parsing of the jwt token easily, hence we
moved to the updated version for the application.

Added Db support for verifying user

- Added code to support verification of the user in the db. However,
there is more work that is remaining on this front. We also need to
update the db layer for new field that can be used to verify the user.

- Added handler code to support user verification and also updated
the service layer for user validation and validation.

Tech debt changes

- Fixed unit test that was erorring out after moving to new version
for the jwt pacakge ciri-stormrider

* Add db layer support to build user verification

Added check in db layer to add user verification. Once the user
is verified, the verification check is made and the user can now
log into the application with ease.

* Update DB and Web Layer for Email Service

Created new sql file to handle verified user

- Added new sql file to support the verification of a selected user.
Also updated test to support verified users. Currently, both test users
are labelled as verified, but when the MVP is rolled out, and new changes
are sought out for the implications of verified user, we can either
create a new unverified user or label one of the other users as an unverified
user.

- Updated api layer code to support revalidation of email for each user

The verified user route is a little complex since there is multiple avenues
of approach for this situation. Currently, users are able revalidate once
if the email didn't reach them but we want to ensure that the user has the
correct token and authorization so that they can perform the email notification
service again.

Updated model struct and test cases around user service.

- Updated UI to support email re-validation

The ui part for this request is a bit tricky. To prevent the users from
pressing the button again and again, we added a timeout for 60 seconds before
the user can send the email again. However, if the user refreshes the page, the
timer is invalidated. Hence, we needed a better solution. For now, we
simply store the key in the local storage and the next time the user logs
in they are able to request a re-validation again. Currently, there is
no visible / tangible issue with user invalidation. So for MVP we are going
to keep it at this.

* 329 - Update dev script to support token (#331)

- Updated dev script to support custom token

Added .env file changes so that now all jwt tokens
are custom for the server they are under. Users signed with the selected
token must use the same token to decode the key.

Closes #329