335 - Add extra fields to bulk upload
Added extra fields that can be supported by bulk upload. The fields
barcode and sku should not be of number type or else the values inserted
would rather be of type number. In the future we should update the sku and
barcode to be explicitly of type string but currently the assumption is that
such data validity is maintained by the client.

Closes #335