342 forgot pwd (#347)
* 342 - Clean up modal components for auth

- Tech debt cleanup for existing modal components

Created consistent form components for sign up and sign in fields. Also
added modal state to properly manage the login / sign up workflow. We do
not need to have two states to manage that, made it simpler and easier
to the eyes.

* Created ForgotPassword component

- Created forgot password text component

This component is used from login form modal where if the user forgot
their password, they could just simply click on this to have an email
sent to their email address.

- Created storybook for forgot password

- Created forgot password form fields and respective storybook

Created forgot password form field component to enter the email address
so that it can be used to send the token to the selected email address.
Currently, the backend implementation is not up, so all changes are just
visual components.

closes #342