342 Integrated reset password workflow
Resetting a password workflow required us to navigate some unknown
with much contention. The workflow cannot use JWT token as JWT can
only be used with logged in users. Since reset password is equal to
the user not being able to log in, we used context of the page to
determine if the user can view the "Forgot Password" workflow or not.

This was also because the workflow sits outside of the navigate context
of react-router - the user is not logged in. Hence, the page cannot
be loaded deterministically. In order to circumvent this, we created a
context provider that navigates pages based on situation of the page. Once
the user presses "Forgot Password" then the context allows for said workflow
to be displayed during render.